requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
alembic==1.13.2
pydantic[email]==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
mysqlclient==2.2.4
python-multipart==0.0.9